requirements.txt
Django==4.2.3
django-cors-headers==3.13.0
django-environ==0.10.0
django-filter==21.1
djangorestframework==3.14.0
gunicorn==20.1.0
psycopg2-binary==2.9.5
py-asciimath==0.3.0
pytest==7.4.0
pytest-django==4.5.2